Update README and tests for adaptive dynamics
- Enhanced README.md with a new composability contract section, detailing the roles and examples of various layers in the adaptive dynamics.
- Updated the `learning` package to include `NewClassifierWeights` for better handling of predicted vs. actual calibration.
- Refactored tests for `Accumulator`, `Compression`, `Delta`, `EMA`, `Momentum`, `Range`, and `Variance` to improve clarity and ensure they validate the new dynamic constructors.
- Removed outdated tests and benchmarks for state observation methods, streamlining the testing process.
- Added new tests for classifier weights to ensure proper error handling and feature scaling.

+++ statistic/min_test.go
@@ -12,7 +12,7 @@ func TestMin_Observe(testingTB *testing.T) {
 		value := NewMin().Observe(nomagique.Numbers(3, 1, 4, 2)...)
 
 		Convey("It should return the smallest value", func() {
-			So(float64(value), ShouldEqual, 1)
+			So(value, ShouldEqual, 1)
 		})
 	})
 }
